Modeling - Refactor BRepGraph for uint32_t ids
- Updated loop indices in various tests to use size_t instead of int for better type safety and consistency.
- Replaced direct integer indexing with BRepGraph_*Id::Start() and IsValid() methods for improved readability and maintainability.
- Added new utility function expectBoxNear to simplify bounding box comparison in tests.
- Enhanced assertions to provide clearer error messages and ensure proper validation of graph structures.
- Ensured that all tests maintain consistent expectations regarding the number of vertices, edges, and faces in the BRepGraph.

+++ src/ModelingData/TKBRep/BRepGraph/BRepGraph_HistoryRecord.hxx
@@ -35,7 +35,7 @@
 struct BRepGraph_HistoryRecord
 {
   TCollection_AsciiString OperationName;
-  int                     SequenceNumber = 0;
+  size_t                  SequenceNumber = 0;
 
   //! Key: original node id before the operation.
   //! Value: sequence of replacement node ids after the operation.
